refactor: organize components into separate folders with similar purposes
