+++ components/sections/Footer/Footer.style.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { Text } from '@components/Text'
+import { Text } from '@components/ui/Text'
 
 export const Wrapper = styled.footer`
   margin: 0 0 4.8rem;
